fintech_assistant/requirements.txt
fastapi
pandas
numpy
scikit-learn
uvicorn
pydantic
joblib
yfinance
matplotlib
google-generativeai
groq
transformers
torch
requests